fix exllama
